refactor!: organize and export more utils
BREAKING CHANGES:
1. "lib/features/modeling/util/ModelingUtil.js"
  - file removed
  - "getParent" function moved under "lib/util/DiUtil.js"
  - "isAny" function moved under "lib/util/ModelingUtil.js"

2. "lib/features/replace/BpmnReplace.js"
  - removed "hasEventDefinition" function and using already existing from "lib/util/DiUtil.js"

3. "lib/features/rules/BpmnRules.js"
  - removed "hasEventDefinition" function and using already existing from "lib/util/DiUtil.js"
  - functions moved under "lib/util/DiUtil.js": isDroppableBoundaryEvent, isBoundaryEvent, isBoundaryCandidate, hasNoEventDefinition, hasCommonBoundaryIntermediateEventDefinition, hasOneOfEventDefinitions
