+++ lib/features/auto-place/AutoPlaceUtil.js
@@ -1,5 +1,7 @@
-import { is } from '../../util/ModelUtil';
-import { isAny } from '../modeling/util/ModelingUtil';
+import {
+  is,
+  isAny
+} from '../../util/ModelUtil';
 
 import {
   getMid,
