+++ lib/features/context-pad/ContextPadProvider.js
@@ -5,17 +5,14 @@ import {
 } from 'min-dash';
 
 import {
-  is
+  is,
+  isAny
 } from '../../util/ModelUtil';
 
 import {
   isExpanded,
   isEventSubProcess
 } from '../../util/DiUtil';
-
-import {
-  isAny
-} from '../modeling/util/ModelingUtil';
 
 import {
   getChildLanes
