+++ lib/features/distribute-elements/BpmnDistributeElements.js
@@ -4,7 +4,7 @@ import {
 
 import {
   isAny
-} from '../modeling/util/ModelingUtil';
+} from '../../util/ModelUtil';
 
 
 /**
