+++ lib/features/grid-snapping/BpmnGridSnapping.js
@@ -1,4 +1,4 @@
-import { isAny } from '../modeling/util/ModelingUtil';
+import { isAny } from '../../util/ModelUtil';
 
 export default function BpmnGridSnapping(eventBus) {
   eventBus.on([
