+++ lib/features/label-editing/LabelEditingProvider.js
@@ -8,14 +8,14 @@ import {
 
 import {
   getBusinessObject,
-  is
+  is,
+  isAny
 } from '../../util/ModelUtil';
 
 import {
   createCategoryValue
 } from '../modeling/behavior/util/CategoryUtil';
 
-import { isAny } from '../modeling/util/ModelingUtil';
 import { isExpanded } from '../../util/DiUtil';
 
 import {
