+++ lib/features/modeling/BpmnFactory.js
@@ -6,7 +6,7 @@ import {
 
 import {
   isAny
-} from './util/ModelingUtil';
+} from '../../util/ModelUtil';
 
 
 export default function BpmnFactory(moddle) {
