+++ lib/features/modeling/BpmnUpdater.js
@@ -16,12 +16,9 @@ import {
 
 import {
   getBusinessObject,
-  is
-} from '../../util/ModelUtil';
-
-import {
+  is,
   isAny
-} from './util/ModelingUtil';
+} from '../../util/ModelUtil';
 
 import CommandInterceptor from 'diagram-js/lib/command/CommandInterceptor';
