+++ lib/features/modeling/behavior/AttachEventBehavior.js
@@ -2,7 +2,7 @@ import inherits from 'inherits';
 
 import CommandInterceptor from 'diagram-js/lib/command/CommandInterceptor';
 
-import { isAny } from '../util/ModelingUtil';
+import { isAny } from '../../../util/ModelUtil';
 import { getBusinessObject } from '../../../util/ModelUtil';
