+++ lib/features/modeling/behavior/CreateBehavior.js
@@ -1,10 +1,14 @@
 import inherits from 'inherits';
 
-import { is } from '../../../util/ModelUtil';
+import {
+  is
+} from '../../../util/ModelUtil';
+
+import {
+  getParent
+} from '../../../util/DiUtil';
 
 import CommandInterceptor from 'diagram-js/lib/command/CommandInterceptor';
-
-import { getParent } from '../util/ModelingUtil';
 
 
 export default function CreateBehavior(injector) {
