+++ lib/features/modeling/behavior/CreateBoundaryEventBehavior.js
@@ -2,7 +2,7 @@ import inherits from 'inherits';
 
 import CommandInterceptor from 'diagram-js/lib/command/CommandInterceptor';
 
-import { isAny } from '../util/ModelingUtil';
+import { isAny } from '../../../util/ModelUtil';
 import { isLabel } from '../../../util/LabelUtil';
 
 /**
