+++ lib/features/modeling/behavior/DataStoreBehavior.js
@@ -2,9 +2,10 @@ import inherits from 'inherits';
 
 import CommandInterceptor from 'diagram-js/lib/command/CommandInterceptor';
 
-import { is } from '../../../util/ModelUtil';
-
-import { isAny } from '../util/ModelingUtil';
+import {
+  is,
+  isAny
+} from '../../../util/ModelUtil';
 
 import UpdateSemanticParentHandler from '../cmd/UpdateSemanticParentHandler';
