+++ lib/features/modeling/behavior/FixHoverBehavior.js
@@ -1,8 +1,9 @@
 import { getLanesRoot } from '../util/LaneUtil';
 
-import { is } from '../../../util/ModelUtil';
-
-import { isAny } from '../util/ModelingUtil';
+import {
+  is,
+  isAny
+} from '../../../util/ModelUtil';
 
 var HIGH_PRIORITY = 1500;
 var HIGHEST_PRIORITY = 2000;
