+++ lib/features/modeling/behavior/IsHorizontalFix.js
@@ -3,12 +3,9 @@ import inherits from 'inherits';
 import CommandInterceptor from 'diagram-js/lib/command/CommandInterceptor';
 
 import {
+  isAny,
   getBusinessObject
 } from '../../../util/ModelUtil';
-
-import {
-  isAny
-} from '../util/ModelingUtil';
 
 /**
  * A component that makes sure that each created or updated
