+++ lib/features/modeling/util/LaneUtil.js
@@ -1,8 +1,10 @@
-import { is } from '../../../util/ModelUtil';
+import {
+  is
+} from '../../../util/ModelUtil';
 
 import {
   getParent
-} from './ModelingUtil';
+} from '../../../util/DiUtil';
 
 import {
   asTRBL
