+++ lib/features/ordering/BpmnOrderingProvider.js
@@ -4,7 +4,7 @@ import OrderingProvider from 'diagram-js/lib/features/ordering/OrderingProvider'
 
 import {
   isAny
-} from '../modeling/util/ModelingUtil';
+} from '../../util/ModelUtil';
 
 import {
   findIndex,
