+++ lib/features/replace/BpmnReplace.js
@@ -10,14 +10,11 @@ import {
 
 import {
   is,
-  getBusinessObject
+  isAny
 } from '../../util/ModelUtil';
 
 import {
-  isAny
-} from '../modeling/util/ModelingUtil';
-
-import {
+  hasEventDefinition,
   isExpanded,
   isEventSubProcess
 } from '../../util/DiUtil';
@@ -286,15 +283,6 @@ BpmnReplace.$inject = [
 
 function isSubProcess(bo) {
   return is(bo, 'bpmn:SubProcess');
-}
-
-function hasEventDefinition(element, type) {
-
-  var bo = getBusinessObject(element);
-
-  return type && bo.get('eventDefinitions').some(function(definition) {
-    return is(definition, type);
-  });
 }
 
 /**
