+++ lib/features/rules/BpmnRules.js
@@ -9,17 +9,19 @@ import inherits from 'inherits';
 
 import {
   is,
+  isAny,
   getBusinessObject
 } from '../../util/ModelUtil';
 
 import {
-  getParent,
-  isAny
-} from '../modeling/util/ModelingUtil';
-
-import {
   isLabel
 } from '../../util/LabelUtil';
+
+import {
+  hasEventDefinition,
+  isBoundaryCandidate,
+  isDroppableBoundaryEvent
+} from '../../util/DiUtil';
 
 import {
   isExpanded,
@@ -27,7 +29,8 @@ import {
   isInterrupting,
   hasErrorEventDefinition,
   hasEscalationEventDefinition,
-  hasCompensateEventDefinition
+  hasCompensateEventDefinition,
+  getParent
 } from '../../util/DiUtil';
 
 import RuleProvider from 'diagram-js/lib/features/rules/RuleProvider';
@@ -322,14 +325,6 @@ function isSameScope(a, b) {
   return scopeParentA === scopeParentB;
 }
 
-function hasEventDefinition(element, eventDefinition) {
-  var bo = getBusinessObject(element);
-
-  return !!find(bo.eventDefinitions || [], function(definition) {
-    return is(definition, eventDefinition);
-  });
-}
-
 function hasEventDefinitionOrNone(element, eventDefinition) {
   var bo = getBusinessObject(element);
 
@@ -515,58 +510,8 @@ function canDrop(element, target, position) {
   return false;
 }
 
-function isDroppableBoundaryEvent(event) {
-  return getBusinessObject(event).cancelActivity && (
-    hasNoEventDefinition(event) || hasCommonBoundaryIntermediateEventDefinition(event)
-  );
-}
-
-function isBoundaryEvent(element) {
-  return !isLabel(element) && is(element, 'bpmn:BoundaryEvent');
-}
-
 function isLane(element) {
   return is(element, 'bpmn:Lane');
-}
-
-/**
- * We treat IntermediateThrowEvents as boundary events during create,
- * this must be reflected in the rules.
- */
-function isBoundaryCandidate(element) {
-  if (isBoundaryEvent(element)) {
-    return true;
-  }
-
-  if (is(element, 'bpmn:IntermediateThrowEvent') && hasNoEventDefinition(element)) {
-    return true;
-  }
-
-  return (
-    is(element, 'bpmn:IntermediateCatchEvent') &&
-    hasCommonBoundaryIntermediateEventDefinition(element)
-  );
-}
-
-function hasNoEventDefinition(element) {
-  var bo = getBusinessObject(element);
-
-  return bo && !(bo.eventDefinitions && bo.eventDefinitions.length);
-}
-
-function hasCommonBoundaryIntermediateEventDefinition(element) {
-  return hasOneOfEventDefinitions(element, [
-    'bpmn:MessageEventDefinition',
-    'bpmn:TimerEventDefinition',
-    'bpmn:SignalEventDefinition',
-    'bpmn:ConditionalEventDefinition'
-  ]);
-}
-
-function hasOneOfEventDefinitions(element, eventDefinitions) {
-  return eventDefinitions.some(function(definition) {
-    return hasEventDefinition(element, definition);
-  });
 }
 
 function isReceiveTaskAfterEventBasedGateway(element) {
