+++ lib/features/snapping/BpmnConnectSnapping.js
@@ -9,13 +9,14 @@ import {
   getOrientation
 } from 'diagram-js/lib/layout/LayoutUtil';
 
-import { is } from '../../util/ModelUtil';
+import {
+  is,
+  isAny
+} from '../../util/ModelUtil';
 
 import {
   some
 } from 'min-dash';
-
-import { isAny } from '../modeling/util/ModelingUtil';
 
 var HIGHER_PRIORITY = 1250;
