+++ lib/util/DiUtil.js
@@ -1,7 +1,12 @@
 import {
   is,
+  isAny,
   getBusinessObject
 } from './ModelUtil';
+
+import {
+  isLabel
+} from './LabelUtil';
 
 import {
   forEach
@@ -48,6 +53,12 @@ export function hasEventDefinition(element, eventType) {
   return hasEventDefinition;
 }
 
+export function hasNoEventDefinition(element) {
+  var bo = getBusinessObject(element);
+
+  return bo && !(bo.eventDefinitions && bo.eventDefinitions.length);
+}
+
 export function hasErrorEventDefinition(element) {
   return hasEventDefinition(element, 'bpmn:ErrorEventDefinition');
 }
@@ -58,4 +69,71 @@ export function hasEscalationEventDefinition(element) {
 
 export function hasCompensateEventDefinition(element) {
   return hasEventDefinition(element, 'bpmn:CompensateEventDefinition');
+}
+
+function hasCommonBoundaryIntermediateEventDefinition(element) {
+  return hasOneOfEventDefinitions(element, [
+    'bpmn:MessageEventDefinition',
+    'bpmn:TimerEventDefinition',
+    'bpmn:SignalEventDefinition',
+    'bpmn:ConditionalEventDefinition'
+  ]);
+}
+
+function hasOneOfEventDefinitions(element, eventDefinitions) {
+  return eventDefinitions.some(function(definition) {
+    return hasEventDefinition(element, definition);
+  });
+}
+
+/**
+ * Return the parent of the element with any of the given types.
+ *
+ * @param {djs.model.Base} element
+ * @param {String|Array<String>} anyType
+ *
+ * @return {djs.model.Base}
+ */
+export function getParent(element, anyType) {
+
+  if (typeof anyType === 'string') {
+    anyType = [ anyType ];
+  }
+
+  while ((element = element.parent)) {
+    if (isAny(element, anyType)) {
+      return element;
+    }
+  }
+
+  return null;
+}
+
+/**
+ * We treat IntermediateThrowEvents as boundary events during create,
+ * this must be reflected in the rules.
+ */
+export function isBoundaryCandidate(element) {
+  if (isBoundaryEvent(element)) {
+    return true;
+  }
+
+  if (is(element, 'bpmn:IntermediateThrowEvent') && hasNoEventDefinition(element)) {
+    return true;
+  }
+
+  return (
+    is(element, 'bpmn:IntermediateCatchEvent') &&
+    hasCommonBoundaryIntermediateEventDefinition(element)
+  );
+}
+
+export function isBoundaryEvent(element) {
+  return !isLabel(element) && is(element, 'bpmn:BoundaryEvent');
+}
+
+export function isDroppableBoundaryEvent(event) {
+  return getBusinessObject(event).cancelActivity && (
+    hasNoEventDefinition(event) || hasCommonBoundaryIntermediateEventDefinition(event)
+  );
 }
