+++ lib/util/ModelUtil.js
@@ -1,3 +1,8 @@
+import {
+  some
+} from 'min-dash';
+
+
 /**
  * Is an element of the given BPMN type?
  *
@@ -10,6 +15,21 @@ export function is(element, type) {
   var bo = getBusinessObject(element);
 
   return bo && (typeof bo.$instanceOf === 'function') && bo.$instanceOf(type);
+}
+
+
+/**
+ * Return true if element has any of the given types.
+ *
+ * @param {djs.model.Base} element
+ * @param {Array<String>} types
+ *
+ * @return {Boolean}
+ */
+export function isAny(element, types) {
+  return some(types, function(t) {
+    return is(element, t);
+  });
 }
