+++ test/spec/features/editor-actions/BpmnEditorActionsSpec.js
@@ -7,7 +7,7 @@ var pick = require('min-dash').pick;
 
 var getBBox = require('diagram-js/lib/util/Elements').getBBox;
 
-var getParent = require('lib/features/modeling/util/ModelingUtil').getParent;
+var getParent = require('lib/util/DiUtil').getParent;
 
 
 import bpmnEditorActionsModule from 'lib/features/editor-actions';
